pyproject.toml
[project]
name = "tracksensei-telemetry"
version = "0.1.0"
description = "Telemetry app for recording GPS data"
authors = [
    { name = "Rudy Suarez", email = "rudysuarez1@example.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "pyserial (>=3.5,<4.0)",
    "gpsd-py3 (>=0.3.0,<1.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pynmea2 (>=1.19.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "app" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
pre-commit = "^4.2.0"
pyright = "^1.1.398"
